requirement_new.txt
absl-py==0.9.0
aiohttp==2.3.10
akro==0.0.8
astor==0.8.1
async-timeout==2.0.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
blessings==1.7
boto3==1.12.3
botocore==1.15.3
Box2D==2.3.1
Box2D-kengz==2.3.3
box2d-py==2.3.8
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cma==2.7.0
colorama==0.4.3
cycler==0.10.0
Cython==0.28.5
decorator==4.4.1
docutils==0.15.2
dowel==0.0.3
filelock==3.0.12
Flask==1.0.2
future==0.18.2
-e git+https://github.com/rlworkgroup/garage.git@f5daf56653689957830f8969cb9b3fcb7595bfce#egg=garage
gast==0.3.3
gitdb2==3.0.2
GitPython==2.1.7
glfw==1.10.1
google==2.0.3
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.29.0
gtimer==1.0.0b5
gym==0.15.4
h5py==2.10.0
idna==2.9
idna-ssl==1.1.0
imageio==2.8.0
importlib-metadata==1.6.0
ipdb==0.12.3
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.12.1
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.12.5
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.20.0
lockfile==0.12.2
Mako==1.0.6
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==2.0.2
-e git+https://github.com/rlworkgroup/metaworld.git@2f228c261c9592dca1f3b8b89fd633d045862af0#egg=metaworld
mpi4py==3.0.3
msgpack==0.6.2
mujoco-py==2.0.2.5
multidict==4.7.6
networkx==2.4
numba==0.35.0
numpy==1.14.5
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.32
pandas==0.24.2
parso==0.6.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==7.0.0
plotly==4.0.0
prompt-toolkit==1.0.15
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py-spy==0.3.3
pycosat==0.6.3
pycparser==2.19
pygame==1.9.2
pyglet==1.3.2
Pygments==2.5.2
PyOpenGL==3.1.5
pyparsing==2.4.6
PyPrind==2.11.2
pyrsistent==0.16.0
python-dateutil==2.6.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
ray==0.8.5
redis==3.4.1
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-image==0.15.0
scikit-video==1.1.11
scipy==1.0.1
seaborn==0.9.1
setproctitle==1.1.10
simplegeneric==0.8.1
sip==4.19.12
six==1.11.0
smmap2==2.0.5
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==1.14.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-probability==0.10.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.1
yarl==1.4.2
zipp==1.2.0
